Revert "Water gun testing"
This reverts commit e2716d91e994d0007d6318d2ad3e750452862844.

+++ Assets/Scripts/Buildings/BuildingHealth.cs
@@ -14,7 +14,7 @@ namespace OperationPlayground.Buildings
         [System.NonSerialized]
         public BuildingScriptableObject buildingSo;
 
-        public override float MaxHealth => buildingSo.health;
+        public override int MaxHealth => buildingSo.health;
 
         protected override bool DoSpawnHealthBar => false;
